Update `setup.py` and add `requirements.txt`
Changes to numpy and pandas since this code was originally written cause a
`ValueError`, breaking an `assert` in the model comparison code.  Appears due to
numpy beginning to delegate to pandas for `.any()` (and other functions), which
itself returns a Series instead of single boolean, resulting in error "Truth
value of a Series is ambiguous".  Pin numpy and pandas at the versions which
work.

Also add missing xlrd dependency, required during model training.

Create `requirements.txt` for pip to give absolute dependency clarity and
simplify installation.